requirements.txt
grpcio==1.68.1
grpcio-tools==1.68.1
Pillow==11.0.0
protobuf==5.29.1
torch==2.5.1
transformers==4.46.3
